Header: fix mobile navigation
