+++ client/src/app/[locale]/not-found.tsx
@@ -1,7 +1,8 @@
+import { useTranslations } from "next-intl";
+
 import { Button } from "@/components/ui/button";
 
 import { Link } from "@/i18n/navigation";
-import { useTranslations } from "next-intl";
 
 export default function Custom404() {
   const t = useTranslations();
